Automatically backed up by Learn

+++ lib/create.sql
@@ -9,7 +9,7 @@ CREATE TABLE projects (
   id INTEGER PRIMARY KEY, 
   title TEXT, 
   category TEXT, 
-  funding_goal REAL, 
+  funding_goal INTEGER, 
   start_date BLOB, 
   end_date BLOB);
